requirements.txt
Django~=4.0.5
psycopg2-binary~=2.9.3
pyyaml
django-crispy-forms~=1.14.0
djangorestframework~=3.13.1
